components: move badge to components/core
