+++ static/app/components/replays/replayCountBadge.tsx
@@ -1,4 +1,4 @@
-import Badge from 'sentry/components/badge/badge';
+import Badge from 'sentry/components/core/badge';
 
 function ReplayCountBadge({count}: {count: undefined | number}) {
   const display = count && count > 50 ? '50+' : count ?? null;
