+++ static/app/views/alerts/list/rules/teamFilter.tsx
@@ -4,8 +4,8 @@ import debounce from 'lodash/debounce';
 import partition from 'lodash/partition';
 
 import TeamAvatar from 'sentry/components/avatar/teamAvatar';
-import Badge from 'sentry/components/badge/badge';
 import {CompactSelect} from 'sentry/components/compactSelect';
+import Badge from 'sentry/components/core/badge';
 import {DEFAULT_DEBOUNCE_DURATION} from 'sentry/constants';
 import {IconUser} from 'sentry/icons';
 import {t} from 'sentry/locale';
