+++ static/app/views/dashboards/widgets/common/widgetFrame.tsx
@@ -1,9 +1,9 @@
 import {Fragment} from 'react';
 import styled from '@emotion/styled';
 
-import type {BadgeProps} from 'sentry/components/badge/badge';
-import Badge from 'sentry/components/badge/badge';
 import {Button, LinkButton} from 'sentry/components/button';
+import type {BadgeProps} from 'sentry/components/core/badge';
+import Badge from 'sentry/components/core/badge';
 import {DropdownMenu, type MenuItemProps} from 'sentry/components/dropdownMenu';
 import {Tooltip} from 'sentry/components/tooltip';
 import {IconEllipsis, IconExpand, IconWarning} from 'sentry/icons';
