+++ static/app/views/settings/components/settingsNavItem.tsx
@@ -2,8 +2,8 @@ import type {ReactElement} from 'react';
 import {Fragment} from 'react';
 import styled from '@emotion/styled';
 
-import Badge from 'sentry/components/badge/badge';
 import FeatureBadge from 'sentry/components/badge/featureBadge';
+import Badge from 'sentry/components/core/badge';
 import HookOrDefault from 'sentry/components/hookOrDefault';
 import {SecondaryNav} from 'sentry/components/nav/secondary';
 import {Tooltip} from 'sentry/components/tooltip';
