+++ static/app/views/settings/project/navigationConfiguration.tsx
@@ -1,4 +1,4 @@
-import Badge from 'sentry/components/badge/badge';
+import Badge from 'sentry/components/core/badge';
 import {t} from 'sentry/locale';
 import ConfigStore from 'sentry/stores/configStore';
 import type {Organization} from 'sentry/types/organization';
